@@ -1,7 +1,7 @@
 [project]
 name = "metatrain"
 dynamic = ["version"]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 
 readme = "README.md"
 license = "BSD-3-Clause"
@@ -14,8 +14,8 @@ dependencies = [
     "huggingface_hub",
     "metatensor-learn >=0.3.2,<0.4",
     "metatensor-operations >=0.3.4,<0.4",
-    "metatensor-torch >=0.7.6,<0.8",
-    "metatomic-torch >=0.1.2,<0.2",
+    "metatensor-torch >=0.8.1,<0.9",
+    "metatomic-torch >=0.1.5,<0.2",
     "jsonschema",
     "omegaconf >= 2.3.0",
     "python-hostlist",
@@ -65,14 +65,6 @@ soap-bpnn = [
     "torch-spex>=0.1,<0.2",
     "wigners",
 ]
-deprecated-pet = [
-    "pet-neighbors-convert",
-    "torch_geometric",
-    "matscipy",
-    "scikit-learn",
-    "pathos",
-    "scipy",
-]
 pet = []
 nanopet = []
 gap = [
@@ -80,6 +72,7 @@ gap = [
     "skmatter",
     "scipy",
 ]
+llpr = []
 
 [tool.check-manifest]
 ignore = ["src/metatrain/_version.py"]
@@ -145,9 +138,11 @@ filterwarnings = [
     "ignore:Second derivatives of the spherical harmonics with respect to the Cartesian coordinates were not requested at class creation.",
     # OldCompositionModel deprecation warning
     "ignore:`OldCompositionModel` composition model is deprecated. Please use `CompositionModel` instead.",
-    # We don't yet have a linux-aarch64 wheel on PyPI, so libgomp is not copied to 
+    # We don't yet have a linux-aarch64 wheel on PyPI, so libgomp is not copied to
     # sphericart_torch.libs' by cibuildwheels
     "ignore:No libgomp shared library found in 'sphericart_torch.libs'.",
+    # Multi-threaded tests clash with multi-process data-loading
+    "ignore:This process \\(pid=\\d+\\) is multi-threaded, use of fork\\(\\) may lead to deadlocks in the child.:DeprecationWarning",
 ]
 
 [tool.yamlfix]
