@@ -8,7 +8,12 @@ from metatensor.torch import Labels, TensorBlock, TensorMap
 ###########################
 
 
-def model_update_v1_v2(checkpoint):
+def model_update_v1_v2(checkpoint: dict) -> None:
+    """
+    Update a v1 checkpoint to v2.
+
+    :param checkpoint: The checkpoint to update.
+    """
     for key in ["model_state_dict", "best_model_state_dict"]:
         if (state_dict := checkpoint.get(key)) is not None:
             if "additive_models.0.model.type_to_index" not in state_dict:
@@ -17,7 +22,12 @@ def model_update_v1_v2(checkpoint):
                 )
 
 
-def model_update_v2_v3(checkpoint):
+def model_update_v2_v3(checkpoint: dict) -> None:
+    """
+    Update a v2 checkpoint to v3.
+
+    :param checkpoint: The checkpoint to update.
+    """
     for key in ["model_state_dict", "best_model_state_dict"]:
         if (state_dict := checkpoint.get(key)) is not None:
             if "train_hypers" in state_dict:
@@ -27,7 +37,12 @@ def model_update_v2_v3(checkpoint):
             state_dict["finetune_config"] = finetune_config
 
 
-def model_update_v3_v4(checkpoint):
+def model_update_v3_v4(checkpoint: dict) -> None:
+    """
+    Update a v3 checkpoint to v4.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["epoch"] = checkpoint.get("epoch")
     checkpoint["best_epoch"] = checkpoint.get("best_epoch")
 
@@ -35,18 +50,33 @@ def model_update_v3_v4(checkpoint):
         checkpoint["best_model_state_dict"] = checkpoint.get("best_model_state_dict")
 
 
-def model_update_v4_v5(checkpoint):
+def model_update_v4_v5(checkpoint: dict) -> None:
+    """
+    Update a v4 checkpoint to v5.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["model_data"]["dataset_info"]._atomic_types = list(
         checkpoint["model_data"]["dataset_info"]._atomic_types
     )
 
 
-def model_update_v5_v6(checkpoint):
+def model_update_v5_v6(checkpoint: dict) -> None:
+    """
+    Update a v5 checkpoint to v6.
+
+    :param checkpoint: The checkpoint to update.
+    """
     if not checkpoint["best_model_state_dict"]:
         checkpoint["best_model_state_dict"] = checkpoint["model_state_dict"]
 
 
-def model_update_v6_v7(checkpoint):
+def model_update_v6_v7(checkpoint: dict) -> None:
+    """
+    Update model checkpoint from version 6 to version 7.
+
+    :param checkpoint: The checkpoint to be updated.
+    """
     # this update consists in changes in the scaler
     for key in ["model_state_dict", "best_model_state_dict"]:
         if (state_dict := checkpoint.get(key)) is not None:
@@ -114,17 +144,32 @@ def model_update_v6_v7(checkpoint):
 ###########################
 
 
-def trainer_update_v1_v2(checkpoint):
+def trainer_update_v1_v2(checkpoint: dict) -> None:
+    """
+    Update a v1 Trainer checkpoint to v2.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["train_hypers"]["scheduler_factor"] = checkpoint["train_hypers"].get(
         "scheduler_factor", 0.5
     )
 
 
-def trainer_update_v2_v3(checkpoint):
+def trainer_update_v2_v3(checkpoint: dict) -> None:
+    """
+    Update a v2 Trainer checkpoint to v3.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["best_epoch"] = checkpoint.get("best_epoch")
 
 
-def trainer_update_v3_v4(checkpoint):
+def trainer_update_v3_v4(checkpoint: dict) -> None:
+    """
+    Update a v3 Trainer checkpoint to v4.
+
+    :param checkpoint: The checkpoint to update.
+    """
     old_loss_hypers = checkpoint["train_hypers"]["loss"].copy()
     dataset_info = checkpoint["model_data"]["dataset_info"]
     new_loss_hypers = {}
@@ -139,12 +184,33 @@ def trainer_update_v3_v4(checkpoint):
     checkpoint["train_hypers"]["loss"] = new_loss_hypers
 
 
-def trainer_update_v4_v5(checkpoint):
+def trainer_update_v4_v5(checkpoint: dict) -> None:
+    """
+    Update a v4 Trainer checkpoint to v5.
+
+    :param checkpoint: The checkpoint to update.
+    """
     raise ValueError(
         "In order to use this checkpoint, you need metatrain 2025.10 or earlier. "
         "You can install it with `pip install metatrain==2025.10`."
     )
 
 
-def trainer_update_v5_v6(checkpoint):
+def trainer_update_v5_v6(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 5 to version 6.
+
+    :param checkpoint: The checkpoint to update.
+    """
+    # num_workers=0 means that the main process will do the data loading, which is
+    # equivalent to not setting it (this was the behavior before v6)
+    checkpoint["train_hypers"]["num_workers"] = 0
+
+
+def trainer_update_v6_v7(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 6 to version 7.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["train_hypers"]["fixed_scaling_weights"] = {}
