@@ -8,7 +8,12 @@ from metatensor.torch import Labels, TensorBlock, TensorMap
 ###########################
 
 
-def model_update_v1_v2(checkpoint):
+def model_update_v1_v2(checkpoint: dict) -> None:
+    """
+    Update model checkpoint from version 1 to version 2.
+
+    :param checkpoint: The checkpoint to update.
+    """
     for key in ["model_state_dict", "best_model_state_dict"]:
         if (state_dict := checkpoint.get(key)) is not None:
             state_dict["additive_models.0.model.type_to_index"] = state_dict.pop(
@@ -16,7 +21,12 @@ def model_update_v1_v2(checkpoint):
             )
 
 
-def model_update_v2_v3(checkpoint):
+def model_update_v2_v3(checkpoint: dict) -> None:
+    """
+    Update model checkpoint from version 2 to version 3.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["epoch"] = checkpoint.get("epoch")
     checkpoint["best_epoch"] = checkpoint.get("best_epoch")
 
@@ -24,7 +34,12 @@ def model_update_v2_v3(checkpoint):
         checkpoint["best_model_state_dict"] = checkpoint.get("best_model_state_dict")
 
 
-def model_update_v3_v4(checkpoint):
+def model_update_v3_v4(checkpoint: dict) -> None:
+    """
+    Update model checkpoint from version 3 to version 4.
+
+    :param checkpoint: The checkpoint to be updated.
+    """
     # this update consists in changes in the scaler
     for key in ["model_state_dict", "best_model_state_dict"]:
         if (state_dict := checkpoint.get(key)) is not None:
@@ -92,11 +107,21 @@ def model_update_v3_v4(checkpoint):
 ###########################
 
 
-def trainer_update_v1_v2(checkpoint):
+def trainer_update_v1_v2(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 1 to version 2.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["best_epoch"] = checkpoint.get("best_epoch")
 
 
-def trainer_update_v2_v3(checkpoint):
+def trainer_update_v2_v3(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 2 to version 3.
+
+    :param checkpoint: The checkpoint to update.
+    """
     old_loss_hypers = checkpoint["train_hypers"]["loss"].copy()
     dataset_info = checkpoint["model_data"]["dataset_info"]
     new_loss_hypers = {}
@@ -111,5 +136,21 @@ def trainer_update_v2_v3(checkpoint):
     checkpoint["train_hypers"]["loss"] = new_loss_hypers
 
 
-def trainer_update_v3_v4(checkpoint):
+def trainer_update_v3_v4(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 3 to version 4.
+
+    :param checkpoint: The checkpoint to update.
+    """
+    # num_workers=0 means that the main process will do the data loading, which is
+    # equivalent to not setting it (this was the behavior before v4)
+    checkpoint["train_hypers"]["num_workers"] = 0
+
+
+def trainer_update_v4_v5(checkpoint: dict) -> None:
+    """
+    Update trainer checkpoint from version 4 to version 5.
+
+    :param checkpoint: The checkpoint to update.
+    """
     checkpoint["train_hypers"]["fixed_scaling_weights"] = {}
