@@ -1,4 +1,3 @@
 from .composition import CompositionModel  # noqa: F401
-from .old_composition import OldCompositionModel  # noqa: F401
-from .remove import remove_additive  # noqa: F401
+from .remove import get_remove_additive_transform, remove_additive  # noqa: F401
 from .zbl import ZBL  # noqa: F401
