@@ -1,5 +1,5 @@
 import random
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional, Tuple, Union
 
 import metatensor.torch as mts
 import numpy as np
@@ -12,13 +12,21 @@ from . import torch_jit_script_unless_coverage
 from .data import TargetInfo
 
 
-def get_random_rotation():
-    """Random 3D rotation that is Haar uniformly distributed over SO(3)."""
+def get_random_rotation() -> Rotation:
+    """
+    Random 3D rotation that is Haar uniformly distributed over SO(3).
+
+    :return: a random 3D rotation
+    """
     return Rotation.random()
 
 
-def get_random_inversion():
-    """Randomly choose an inversion factor -1 or 1."""
+def get_random_inversion() -> int:
+    """
+    Randomly choose an inversion factor -1 or 1.
+
+    :return: either -1 or 1
+    """
     return random.choice([1, -1])
 
 
@@ -30,6 +38,9 @@ class RotationalAugmenter:
     :param target_info_dict: A dictionary mapping target names to their corresponding
         :class:`TargetInfo` objects. This is used to determine the type of targets and
         how to apply the augmentations.
+    :param extra_data_info_dict: An optional dictionary mapping extra data names to
+        their corresponding :py:class:`TargetInfo` objects. This is used to determine
+        the type of extra data and how to apply the augmentations.
     """
 
     def __init__(
@@ -101,11 +112,15 @@ class RotationalAugmenter:
         extra_data: Optional[Dict[str, TensorMap]] = None,
     ) -> Tuple[List[System], Dict[str, TensorMap], Dict[str, TensorMap]]:
         """
-        Apply a random augmentation to a number of ``System`` objects and its targets.
+        Apply a random augmentation to a number of :py:class:`System` objects and its
+        targets.
 
-        :param systems: A list of :class:`System` objects to be augmented.
+        :param systems: A list of :py:class:`System` objects to be augmented.
         :param targets: A dictionary mapping target names to their corresponding
-            :class:`TensorMap` objects. These are the targets to be augmented.
+            :py:class:`TensorMap` objects. These are the targets to be augmented.
+        :param extra_data: An optional dictionary mapping extra data names to their
+            corresponding :py:class:`TensorMap` objects. This extra data will also be
+            augmented if provided.
 
         :return: A tuple containing the augmented systems and targets.
         """
@@ -113,7 +128,8 @@ class RotationalAugmenter:
         rotations = [get_random_rotation() for _ in range(len(systems))]
         inversions = [get_random_inversion() for _ in range(len(systems))]
         transformations = [
-            torch.from_numpy(r.as_matrix() * i) for r, i in zip(rotations, inversions)
+            torch.from_numpy(r.as_matrix() * i)
+            for r, i in zip(rotations, inversions, strict=True)
         ]
 
         wigner_D_matrices = {}
@@ -133,7 +149,9 @@ class RotationalAugmenter:
                 if extra_data is not None
                 else [self.target_info_dict]
             )
-            for tensormap_dict, info_dict in zip(tensormap_dicts, info_dicts):
+            for tensormap_dict, info_dict in zip(
+                tensormap_dicts, info_dicts, strict=True
+            ):
                 for name in tensormap_dict.keys():
                     if name.endswith("_mask"):
                         # skip loss masks
@@ -191,7 +209,7 @@ def _apply_wigner_D_matrices(
         new_values = []
         ell = (len(block.components[0]) - 1) // 2
         for v, transformation, wigner_D_matrix in zip(
-            split_values, transformations, wigner_D_matrices[ell]
+            split_values, transformations, wigner_D_matrices[ell], strict=True
         ):
             is_inverted = torch.det(transformation) < 0
             new_v = v.clone()
@@ -229,7 +247,7 @@ def _apply_random_augmentations(
     # Apply the transformations to the systems
 
     new_systems: List[System] = []
-    for system, transformation in zip(systems, transformations):
+    for system, transformation in zip(systems, transformations, strict=True):
         new_system = System(
             positions=system.positions @ transformation.T,
             types=system.types,
@@ -261,7 +279,7 @@ def _apply_random_augmentations(
             new_extra_data[key] = extra_data.pop(key)
 
     for tensormap_dict, new_dict in zip(
-        [targets, extra_data], [new_targets, new_extra_data]
+        [targets, extra_data], [new_targets, new_extra_data], strict=True
     ):
         if tensormap_dict is None:
             continue
@@ -362,7 +380,9 @@ def _apply_random_augmentations(
                     else:
                         split_vectors = torch.split(vectors, [1 for _ in systems])
                     new_vectors = []
-                    for v, transformation in zip(split_vectors, transformations):
+                    for v, transformation in zip(
+                        split_vectors, transformations, strict=True
+                    ):
                         # fold property dimension in, apply transformation,
                         # unfold property dimension
                         new_v = v.transpose(1, 2)
@@ -392,7 +412,9 @@ def _apply_random_augmentations(
                     else:
                         split_tensors = torch.split(tensor, [1 for _ in systems])
                     new_tensors = []
-                    for tensor, transformation in zip(split_tensors, transformations):
+                    for tensor, transformation in zip(
+                        split_tensors, transformations, strict=True
+                    ):
                         new_tensor = torch.einsum(
                             "Aa,iabp,bB->iABp", transformation, tensor, transformation.T
                         )
@@ -413,7 +435,7 @@ def _apply_random_augmentations(
     return new_systems, new_targets, new_extra_data
 
 
-def _complex_to_real_spherical_harmonics_transform(ell: int):
+def _complex_to_real_spherical_harmonics_transform(ell: int) -> np.ndarray:
     # Generates the transformation matrix from complex spherical harmonics
     # to real spherical harmonics for a given l.
     # Returns a transformation matrix of shape ((2l+1), (2l+1)).
@@ -442,7 +464,9 @@ def _complex_to_real_spherical_harmonics_transform(ell: int):
     return U
 
 
-def _scipy_quaternion_to_quaternionic(q_scipy):
+def _scipy_quaternion_to_quaternionic(
+    q_scipy: Union[np.ndarray, List[float]],
+) -> np.ndarray:
     # This function convert a quaternion obtained from the scipy library to the format
     # used by the quaternionic library.
     # Note: 'xyzw' is the format used by scipy.spatial.transform.Rotation
