@@ -1,4 +1,7 @@
+from typing import Dict
+
 import torch
+from metatomic.torch import ModelOutput
 
 
 class DistributedDataParallel(torch.nn.parallel.DistributedDataParallel):
@@ -8,5 +11,5 @@ class DistributedDataParallel(torch.nn.parallel.DistributedDataParallel):
     and adds a function to retrieve the supported outputs of the module.
     """
 
-    def supported_outputs(self):
+    def supported_outputs(self) -> Dict[str, ModelOutput]:
         return self.module.supported_outputs()
