@@ -1,2 +1,2 @@
-from .remove import remove_scale  # noqa: F401
+from .remove import get_remove_scale_transform, remove_scale  # noqa: F401
 from .scaler import Scaler  # noqa: F401
